chore: unified file downloads in a reusable function
Changelog: None
