+++ src/js/components/auditlogs/eventdetails/terminalplayer.js
@@ -11,6 +11,7 @@ import { DEVICE_MESSAGE_TYPES as MessageTypes, DEVICE_MESSAGE_PROTOCOLS as Messa
 import { CloudDownload, Pause, PlayArrow, Refresh } from '@mui/icons-material';
 import XTerm from '../../common/xterm';
 import { colors } from '../../../themes/Mender';
+import { createFileDownload } from '../../../helpers';
 
 const MessagePack = msgpack5();
 const fitAddon = new FitAddon();
@@ -239,13 +240,7 @@ export const TerminalPlayer = ({ className, item, sessionInitialized }) => {
   const onDownloadClick = () => {
     // eslint-disable-next-line no-undef
     const text = generateHtml({ fit: XTERM_FIT_VERSION, search: XTERM_SEARCH_VERSION, xterm: XTERM_VERSION }, buffer);
-    let link = document.createElement('a');
-    link.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(text));
-    link.setAttribute('download', 'terminalsession.html');
-    link.style.display = 'none';
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
+    createFileDownload(text, 'terminalsession.html');
   };
 
   return (
